Fix compilation warnings

+++ src/s2geography/wkt-reader.cc
@@ -93,7 +93,7 @@ class Parser {
 
   int64_t charsLeftInBuffer() { return this->length - this->offset; }
 
-  bool checkBuffer(int n_chars) {
+  bool checkBuffer(int64_t n_chars) {
     return (this->charsLeftInBuffer() - n_chars) >= 0;
   }
 
@@ -724,7 +724,7 @@ WKTReader::WKTReader(const util::Constructor::Options& options)
 }
 
 std::unique_ptr<Geography> WKTReader::read_feature(const char* text,
-                                                     int64_t size) {
+                                                   int64_t size) {
   constructor_->feat_start();
   reader_->read_buffer(constructor_.get(),
                        reinterpret_cast<const uint8_t*>(text), size);
